Partially working
Jackson and I worked on this for the whole 2 hour meeting ran into a bunch of issues the Y axis is somehow locked in my scene we couldn't figure out how to re-enable it. The one moves down and for some reason won't move up. Going to try and fix it but can't promise it'll be done by Monday.