Justice visuals + tutorial tweaks
Added back the wires and bar of the justice scales. I also added a gold texture to the coins and replaced the pennies with coins to avoid the visuals from clashing. The coin stacks also got new box colliders that should make it easier to jump on them. In the tutorial I adjusted trigger boxes to avoid the player jumping over them, along with tweaking the bridge animation for better clarity.